requirements.txt
openai
pydantic
atproto
rich
python-dotenv
flask